+++ packages/components/src/__rc__/LinkButton/v3/_docs/LinkButton--api-specification.mdx
@@ -2,11 +2,11 @@ import { Canvas, Meta, Controls, ArgTypes, DocsStory } from '@storybook/blocks'
 import { ResourceLinks, KAIOInstallation, LinkTo } from '~storybook/components'
 import * as exampleStories from './LinkButton.doc.stories'
 
-<Meta title="Actions/LinkButton/LinkButton (v3)/API Specification" />
+<Meta title="Components/LinkButton/LinkButton (v3)/API Specification" />
 
 # LinkButton API Specification (v3)
 
-Updated Nov 26, 2024
+Updated Dec 18, 2024
 
 <ResourceLinks
   sourceCode="https://github.com/cultureamp/kaizen-design-system/tree/main/packages/components/src/__actions__/LinkButton/v3"
