+++ packages/components/src/__rc__/LinkButton/v3/_docs/LinkButton--usage-guidelines.mdx
@@ -2,11 +2,11 @@ import { Canvas, Meta, Controls } from '@storybook/blocks'
 import { ResourceLinks, KAIOInstallation, LinkTo } from '~storybook/components'
 import * as LinkButton from './LinkButton.doc.stories'
 
-<Meta title="Actions/LinkButton/LinkButton (v3)/Usage Guidelines" />
+<Meta title="Components/LinkButton/LinkButton (v3)/Usage Guidelines" />
 
 # LinkButton (v3)
 
-Updated Nov 26, 2024
+Updated Dec 18, 2024
 
 <ResourceLinks
   sourceCode="https://github.com/cultureamp/kaizen-design-system/tree/main/packages/components/src/__actions__/Button/v3"
