+++ packages/components/src/__rc__/index.ts
@@ -5,3 +5,4 @@ export * from './Select'
 export * from './Tag'
 export * from './Tabs'
 export * from './Tooltip'
+export * from './LinkButton'
